Add Car vehicle types. Add bug fixes and finalize menu functionality

+++ Program.hpp
@@ -104,23 +104,56 @@ public:
   static void addCar(UnorderedLinkedList<T> *cl, UnorderedLinkedList<T2> *rl)
   {
     string plate, make, model;
+    Car::vehicleType vType;
+    int vehicleTypeChoice;
     double price;
     bool available;
 
     cout << "Please enter all vehicle information:" << endl;
     cout << "License Plate Number: ";
     cin >> plate;
-    cin.ignore();
-    cout << "Make: ";
-    getline(cin, make);
-    cout << "Model: ";
-    getline(cin, model);
-    cout << "Price-per-day: ";
-    cin >> price;
 
-    cl->PutItem(Car(plate, make, model, price));
-    cout << "Car has been added to the list!" << endl;
+    // Check if license plate number exists
+    auto c = new Car(plate);
+    if (cl->GetItem(*c) != nullptr)
+    {
+      cout << "Car with license plate number " << plate << " already exists." << endl;
+    }
+    else
+    {
+      cin.ignore();
+      cout << "Make: ";
+      getline(cin, make);
+      cout << "Model: ";
+      getline(cin, model);
+      cout << "Enter the number of your vehicle type:" << endl;
+      cout << "1. Coupe" << endl;
+      cout << "2. Sedan" << endl;
+      cout << "3. SUV" << endl;
+      cout << "4. Exotic" << endl;
+      cin >> vehicleTypeChoice;
+      switch (vehicleTypeChoice - 1)
+      {
+      case 0:
+        vType = Car::vehicleType::Coupe;
+        break;
+      case 1:
+        vType = Car::vehicleType::Sedan;
+        break;
+      case 2:
+        vType = Car::vehicleType::SUV;
+        break;
+      case 3:
+        vType = Car::vehicleType::Exotic;
+        break;
+      }
+      cout << "Price-per-day: ";
+      cin >> price;
 
+      cl->PutItem(Car(plate, make, model, vType, price));
+      cout << "Car has been added to the list!" << endl;
+    }
+    delete c;
     Program<T, T2>::returnToMenu(cl, rl);
   }
   static void removeCar(UnorderedLinkedList<T> *cl, UnorderedLinkedList<T2> *rl)
@@ -170,16 +203,20 @@ public:
     if (cl->GetItem(*c) != nullptr)
     {
       auto *car = cl->GetItem(*c);
-      car->setAvailable(false);
-
-      rl->PutItem(Reservation(name, plate));
-      cout << "Car with license plate number " << plate << " has successfully been reserved.\n\n";
+      if (car->getAvailability() == false)
+        cout << "Car with license plate number " << plate << " has already been reserved." << endl;
+      else
+      {
+        car->setAvailable(false);
+        rl->PutItem(Reservation(name, plate));
+        cout << "Car with license plate number " << plate << " has successfully been reserved.\n\n";
+      }
     }
     else
     {
       cout << "Car was not found within the list." << endl;
     }
-
+    delete c;
     Program<T, T2>::returnToMenu(cl, rl);
   }
   static void cancelReservation(UnorderedLinkedList<T> *cl, UnorderedLinkedList<T2> *rl)
@@ -199,16 +236,24 @@ public:
     if (rl->GetItem(*r) != nullptr)
     {
       auto *car = cl->GetItem(*c);
-      car->setAvailable(true);
-
-      rl->DeleteItem(*r);
-      cout << "Reservation under " << name << " has successfully been cancelled.\n\n";
+      if (car)
+      {
+        car->setAvailable(true);
+        rl->DeleteItem(*r);
+        cout << "Reservation under " << name << " has successfully been cancelled.\n\n";
+      }
+      else
+      {
+        cout << "Reservation was not found within the list." << endl;
+      }
     }
     else
     {
       cout << "Reservation was not found within the list." << endl;
     }
 
+    delete r;
+    delete c;
     Program<T, T2>::returnToMenu(cl, rl);
   }
 };
